+++ R/util_filter.R
@@ -1,21 +1,29 @@
-filter_results = function(x, ...) {
-  UseMethod("filter_results", x)
+filter_results = function(res, pattern = NULL, exclude = FALSE, col = NULL) {
+  UseMethod("filter_results", res)
 }
 
-filter_results.default = function(res, pattern = NULL, exclude = FALSE) {
-  if (!is.null(pattern)) {
-    subset = grepl(pattern, res)
-    if (exclude) res = res[!subset]
-    else         res = res[subset]
-  }
-  res
+#' @exportS3Method
+filter_results.default = function(res, pattern = NULL, exclude = FALSE, col = NULL) {
+  if (is.null(pattern))
+    return(res)
+
+  subset = grepl(pattern, res)
+  if (exclude)
+    res[!subset]
+  else
+    res[subset]
 }
 
-filter_results.data.frame = function(res, col, pattern = NULL, exclude = FALSE) {
-  if (!is.null(pattern)) {
-    subset = grepl(pattern, res[[col]])
-    if (exclude) res = res[!subset,]
-    else         res = res[subset,]
-  }
-  res
+#' @exportS3Method
+filter_results.data.frame = function(res, pattern = NULL, exclude = FALSE, col = NULL) {
+  if (is.null(pattern))
+    return(res)
+
+  stopifnot(!is.null(col))
+
+  subset = grepl(pattern, res[[col]])
+  if (exclude)
+    res[!subset,]
+  else
+    res[subset,]
 }