+++ NEWS.md
@@ -1,5 +1,7 @@
 # ghclass (development version)
 
+* Added check to `repo_mirror_template()` for empty repositories to avoid cryptic GitHub api error.
+
 # ghclass 0.3.0
 
 * Added support for basig GitHub Pages API endpoints - see `pages_enabled()`, `pages_status()`, `pages_create()`, and `pages_delete()`.