+++ NEWS.md
@@ -1,5 +1,7 @@
 # ghclass (development version)
 
+* Added check to `repo_mirror_template()` for empty repositories to avoid cryptic GitHub api error.
+
 # ghclass 0.3.0
 
 * Added support for basig GitHub Pages API endpoints - see `pages_enabled()`, `pages_status()`, `pages_create()`, and `pages_delete()`.
@@ -23,6 +25,10 @@
 * Added `add_badges` argument to `org_create_assignment()`
 
 * Added support for GitHub's versioned REST API
+
+* Increased the default value of `github_set_api_limit()` to 10,000.
+
+* Any rest API request will now report if the request limit is reached via a warning.
 
 # ghclass 0.2.1